refactor(vta): move step-up gate engine to operations; fix layering inversion (P2.4)
The step-up gate's resolution algorithm lived in
routes/trust_tasks/step_up.rs and messaging::handlers reached *up* into
routes:: to resolve a gate — the inversion operations::step_up_policy
documented. This moves the engine into the operations layer.

- New operations::step_up: the StepUpDecision enum + resolve_step_up (the
  pure policy-floor resolution — already took config + acl_ks, not
  &AppState, since P0.13a) + the `op` op-class module. The two
  resolve_step_up unit tests move with it.
- routes/trust_tasks/step_up.rs keeps the route/transport concerns — the
  RequireStepUp axum extractor, the require_step_up wrapper (decision →
  403/reject + approve-request push), the approve-response Trust Task
  handler, challenge minting, and the did-signed/webauthn gate verification
  — and imports the engine from operations. It re-exports `op` so the
  ~15 `super::step_up::op::*` call sites across the acl/keys/contexts/vault
  slices are unchanged.
- messaging::handlers resolves acl/swap-key step-up via
  crate::operations::step_up::* — no longer importing routes::. Inversion
  resolved.
- operations::step_up_policy doc updated to point at the engine's new home.

Behaviour unchanged. step_up unit tests (incl. the moved engine tests) +
the step_up_policy / step_up_approve_response integration suites green;
fmt + clippy clean (all-targets); tee compiles; full suite (965) green.

Note: the wider P2.4 (backup_blob → operations; dispatch_trust_task_core
typed return so messaging stops importing routes:: for dispatch too) remains.
