+++ vta-service/src/operations/mod.rs
@@ -46,6 +46,7 @@ pub mod protocol;
 #[cfg(feature = "webvh")]
 pub mod provision_integration;
 pub mod seeds;
+pub mod step_up;
 pub mod step_up_approval;
 pub mod step_up_policy;
 pub mod vault;
